Changed directories
changed whole organization of the project: now the model and view are stored in the source folder. There is no controller folder, and instead, view controller and use cases files should all be stored in the same folder specific for each use case. Simple views (the ones that don't need a controller) remain in the base `view` folder.